board permissions

+++ src/main/java/com/project/bcl_back/controller/BoardController.java
@@ -4,7 +4,11 @@ import com.project.bcl_back.common.constants.ApiMappingPattern;
 import com.project.bcl_back.dto.ResponseDto;
 import com.project.bcl_back.dto.board.request.BoardRequestDto;
 import com.project.bcl_back.dto.board.response.BoardResponseDto;
+import com.project.bcl_back.provider.JwtProvider;
+import com.project.bcl_back.repository.MatchRepository;
 import com.project.bcl_back.service.BoardDataService;
+import com.project.bcl_back.service.MatchService;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -21,14 +25,30 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BoardController {
     private final BoardDataService boardDataService;
+    private final MatchService matchService;
+    private final JwtProvider jwtProvider;
+
     // 게시글 등록
     @PostMapping(value="/{categoryId}" ,consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<ResponseDto<BoardResponseDto>> createPost(
             @PathVariable Long matchId,
             @PathVariable Long categoryId,
             @RequestPart("data") @Valid BoardRequestDto dto,
-            @RequestPart(value = "file", required = false) MultipartFile file
+            @RequestPart(value = "file", required = false) MultipartFile file,
+            HttpServletRequest request
     )throws IOException {
+        String authHeader = request.getHeader("Authorization");
+
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        String token = authHeader.substring(7);
+        Long userId = jwtProvider.getUserIdFromJwt(token);
+        if (!matchService.isUserInMatch(userId, matchId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         ResponseDto<BoardResponseDto> response = boardDataService.createPost(dto, file);
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
@@ -37,8 +57,20 @@ public class BoardController {
     @GetMapping("/{categoryId}")
     public ResponseEntity<ResponseDto<List<BoardResponseDto>>> getPostByCategory(
             @PathVariable Long matchId,
-            @PathVariable("categoryId") int categoryId
+            @PathVariable("categoryId") int categoryId,
+            HttpServletRequest request
     ){
+        String authHeader = request.getHeader("Authorization");
+
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        String token = authHeader.substring(7);
+        Long userId = jwtProvider.getUserIdFromJwt(token);
+        if (!matchService.isUserInMatch(userId, matchId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         ResponseDto<List<BoardResponseDto>> posts = boardDataService.getPostByCategory(categoryId);
         return ResponseEntity.ok(posts);
     }
@@ -48,8 +80,20 @@ public class BoardController {
     public ResponseEntity<ResponseDto<BoardResponseDto>> getPost(
             @PathVariable Long matchId,
             @PathVariable("categoryId") int categoryId,
-            @PathVariable Long id
+            @PathVariable Long id,
+            HttpServletRequest request
     ) {
+        String authHeader = request.getHeader("Authorization");
+
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        String token = authHeader.substring(7);
+        Long userId = jwtProvider.getUserIdFromJwt(token);
+        if (!matchService.isUserInMatch(userId, matchId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(boardDataService.getPostById(id));
     }
 
@@ -60,8 +104,15 @@ public class BoardController {
             @PathVariable Long id,
             @PathVariable("categoryId") int categoryId,
             @RequestPart("data") @Valid BoardRequestDto dto,
-            @RequestPart(value = "file", required = false) MultipartFile file
+            @RequestPart(value = "file", required = false) MultipartFile file,
+            HttpServletRequest request
     )throws IOException {
+        String token = request.getHeader("Authorization").substring(7);
+        Long userId = jwtProvider.getUserIdFromJwt(token);
+
+        if (!boardDataService.isPostWriter(userId, id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(boardDataService.updatePost(id, dto, file));
     }
 
@@ -70,8 +121,15 @@ public class BoardController {
     public ResponseEntity<ResponseDto<?>> deletePost(
             @PathVariable Long matchId,
             @PathVariable Long id,
-            @PathVariable("categoryId") int categoryId
+            @PathVariable("categoryId") int categoryId,
+            HttpServletRequest request
     ) {
+        String token = request.getHeader("Authorization").substring(7);
+        Long userId = jwtProvider.getUserIdFromJwt(token);
+
+        if (!boardDataService.isPostWriter(userId, id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(boardDataService.deletePost(id));
     }
 }
